# MEDRISE-MEDICAL-CENTRE-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedRise Medical Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pharmacy/stock:
    get:
      operationId: listPharmacyStock
      tags: [pharmacy]
      summary: List pharmacy stock
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: lowStock
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PharmacyStockItem"
    post:
      operationId: createPharmacyStock
      tags: [pharmacy]
      summary: Add a new drug to stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PharmacyStockInput"
      responses:
        "201":
          description: Stock item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmacyStockItem"
  /pharmacy/stock/{id}:
    patch:
      operationId: updatePharmacyStock
      tags: [pharmacy]
      summary: Update a stock item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PharmacyStockInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmacyStockItem"
        "404":
          description: Not found
    delete:
      operationId: deletePharmacyStock
      tags: [pharmacy]
      summary: Delete a stock item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /pharmacy/dispense:
    post:
      operationId: dispensePharmacy
      tags: [pharmacy]
      summary: Dispense medication from stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispenseInput"
      responses:
        "200":
          description: Dispensed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmacyStockItem"
        "400":
          description: Insufficient stock or validation error
...
  /reports/summary:
    get:
      operationId: getReportsSummary
      tags: [reports]
      summary: HMIS summary report
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: HMIS summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportsSummary"
  /reports/targets:
    get:
      operationId: listHmisTargets
      tags: [reports]
      summary: List all HMIS targets
      parameters:
        - name: periodType
          in: query
          required: false
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
      responses:
        "200":
          description: List of HMIS targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HmisTarget"
    put:
      operationId: upsertHmisTargets
      tags: [reports]
      summary: Upsert HMIS targets (admin/owner/medical_director only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HmisTargetUpsertInput"
      responses:
        "200":
          description: Updated targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HmisTarget"
...
components:
  schemas:
...
    PharmacyStockItem:
      type: object
      required: [id, drugName, quantity, unit, reorderLevel, createdAt, updatedAt]
      properties:
        id:
          type: integer
        drugName:
          type: string
        genericName:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        quantity:
          type: integer
        unit:
          type: string
        reorderLevel:
          type: integer
        expiryDate:
          type: ["string", "null"]
        buyingPrice:
          type: ["string", "null"]
        sellingPrice:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PharmacyStockInput:
      type: object
      required: [drugName, quantity, unit]
      properties:
        drugName:
          type: string
        genericName:
          type: string
        category:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        reorderLevel:
          type: integer
        expiryDate:
          type: string
        buyingPrice:
          type: number
        sellingPrice:
          type: number
        notes:
          type: string
    DispenseInput:
      type: object
      required: [stockId, quantity]
      properties:
        stockId:
          type: integer
        quantity:
          type: integer
        patientId:
          type: integer
        consultationId:
          type: integer
        notes:
          type: string
...
    HmisTarget:
      type: object
      required: [id, metricKey, periodType, targetValue, updatedAt]
      properties:
        id:
          type: integer
        metricKey:
          type: string
        periodType:
          type: string
          enum: [daily, weekly, monthly, yearly]
        targetValue:
          type: integer
        updatedBy:
          type: integer
          nullable: true
        updatedAt:
          type: string
    HmisTargetItem:
      type: object
      required: [metricKey, periodType, targetValue]
      properties:
        metricKey:
          type: string
        periodType:
          type: string
          enum: [daily, weekly, monthly, yearly]
        targetValue:
          type: integer
          minimum: 0
    HmisTargetUpsertInput:
      type: object
      required: [targets]
      properties:
        targets:
          type: array
          items:
            $ref: "#/components/schemas/HmisTargetItem"

    # ── REPORTS SCHEMAS ───────────────────────────────────────────────────
    ReportsSummary:
      type: object
      required: [month, totalPatients, newPatients, totalConsultations, totalAppointments, completedAppointments, totalRevenue, totalLabOrders, completedLabOrders, lowStockDrugs]
      properties:
        month:
          type: string
        totalPatients:
          type: integer
        newPatients:
          type: integer
        totalConsultations:
          type: integer
        totalAppointments:
          type: integer
        completedAppointments:
          type: integer
        totalRevenue:
          type: string
        totalLabOrders:
          type: integer
        completedLabOrders:
          type: integer
        lowStockDrugs:
          type: integer
        staffPresent:
          type: integer
        staffAbsent:
          type: integer